# CopilotPlugin/copilot-plugin/appPackage/apiSpecificationFile/openapi.yaml
openapi: "3.0.4"
info:
  title: SSoTme Agent
  version: "1.0"
  description: Agent that uses the ssotme protocol to assist users in editing their projects' single source of truth
servers:
  - url: http://134.209.173.56:8080/
    description: Development server
paths:
  /mark-base:
    post:
      summary: Submit schema modifications to a SSoT base
      description: Submits schema changes like adding fields, creating entities, or updating existing models. Changes are processed asynchronously by the CLI and applied to the SSoT system.
      operationId: writeSsotBase
      parameters:
        - name: baseId
          in: query
          required: true
          description: The alphanumeric base/project ID to modify
          schema:
            type: string
            example: "default"
      requestBody:
        required: true
        description: The updated SSoT schema content to apply
        content:
          application/json:
            schema:
              type: object
              required:
                - content
              properties:
                content:
                  type: string
                  description: Complete SSoT schema data including entities and metadata (Full JSON string)
                  example: "{\"Object\": {\"property\": \"value\"}}"
                command:
                  type: string
                  description: The linux command to use to modify the JSON content of the base
                  example: "sed -i \"\" \"s/helloworld/hello world/"
      responses:
        '200':
          description: The content was successfully MARKED for changes
          content:
            application/json:
              schema:
                type: object
                required:
                  - msg
                properties:
                  msg:
                    type: string
                    description: The server's text response
                    example: "The operation was successful"
                  baseId:
                    type: string
                    description: The base ID that was modified
                    example: "CLI"
        '400':
          description: Missing baseId parameter
          content:
            application/json:
              schema:
                type: object
                properties:
                  msg:
                    type: string
                    description: Error message
                    example: "Missing baseId parameter"
        '422':
          description: Invalid JSON or missing content field
          content:
            application/json:
              schema:
                type: object
                properties:
                  msg:
                    type: string
                    description: Error message
                    example: "Invalid JSON in request body"
                  baseId:
                    type: string
                    example: "CLI"

  /request-read:
    get:
      summary: Read the current SSoT schema for the specified base
      description: Retrieves the current Single Source of Truth schema including all entities, fields, and metadata for the specified base ID
      operationId: readSsotBase
      parameters:
        - name: baseId
          in: query
          required: true
          description: The alphanumeric base/project ID to read from
          schema:
            type: string
            example: "default"
      responses:
        '200':
          description: Successfully retrieved the SSoT schema data. Expected response contains "data" (json object with entities and metadata) and "baseId" (string).
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    description: Complete SSoT schema data including entities and metadata
                  baseId:
                    type: string
                    description: The base/project ID that was requested
                    example: "customer_base"
        '400':
          description: Missing baseId parameter
          content:
            application/json:
              schema:
                type: object
                properties:
                  msg:
                    type: string
                    description: Error message
                    example: "Missing baseId parameter"
        '408':
          description: Request timeout - CLI did not respond within 30 seconds
          content:
            application/json:
              schema:
                type: object
                properties:
                  msg:
                    type: string
                    description: Error message
                    example: "Timeout waiting for response from CLI!"
        '502':
          description: Bad Gateway - CLI response was invalid or empty
          content:
            application/json:
              schema:
                type: object
                properties:
                  msg:
                    type: string
                    description: Error message
                    example: "Error reading response from CLI!"